Make Perl subclasses more transparent to Python
Python's Qt bindings use some kind of low level introspection to find out
what methods a Python subclass of a Qt class overrides. So in order to have
it call the methods of a Perl 6 subclass of a Qt class, we'd have to hook
into this mechanism. Unfortunately there doesn't seem to be a way to do that.

So instead, I now create a Python class for every Perl 6 class inheriting
from Python classes. This generated class uses the standard Python
inheritance mechanism and has proxy methods for all methods the Perl 6
subclass provides.

This way even PyQt's low level mingling finds all the information it needs
and we can write programs using PyQt in Perl 6 :)