Remove obsolete __getattr__ method
We don't need a Perl 6 __getattr__ anymore, since we now create a Python
class with proxies for all the subclass' methods. Thus Python already
knows all available methods including accessors for Perl 6 attributes.

Plus, __getattr__ was broken anyway since it was passed a Buf and it
used it as Str.