Fix segfaults caused by py_unicode_to_char_star
PyString_AsString returns a pointer to the PyString's buffer. So when we
Py_DECREF the string immediately afterwards, this buffer is already free'd.
It still works, but is not reliable.

Instead, move more of the processing to Perl 6, so we can safely py_dec_ref
the string after NativeCall copied it to a Perl 6 string.

Fixes htmlify.p6 in perl6/doc.git